requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
langchain==0.0.174
langchain-community==0.0.14
langchain-experimental==0.0.8
openai==0.27.0  

# chromadb==0.3.21  # Comentar o eliminar si no es necesario
# tiktoken==0.3.0  # Comentar o eliminar si no es necesario
# psycopg2==2.9.3  # Comentar o eliminar si no usas PostgreSQL
# flask-cors==3.0.10  # Comentar o eliminar si no usas CORS
# pymysql==1.0.2  # Comentar o eliminar si no usas MySQL
# sqlalchemy==2.0.0  # Comentar o eliminar si no es necesario
